+++ pycobertura/cobertura.py
@@ -46,15 +46,15 @@ class Cobertura(object):
         self.xml = ET.parse(xml_path).getroot()
 
     @memoize
-    def _get_element_by_class_filename(self, class_filename):
+    def _get_class_element_by_filename(self, filename):
         syntax = "./packages/package/classes/class[@filename='%s'][1]" % (
-            class_filename
+            filename
         )
         return self.xml.xpath(syntax)[0]
 
     @memoize
-    def _get_lines_by_class_filename(self, class_filename):
-        el = self._get_element_by_class_filename(class_filename)
+    def _get_lines_by_filename(self, filename):
+        el = self._get_class_element_by_filename(filename)
         return el.xpath('./lines/line')
 
     @property
@@ -62,58 +62,58 @@ class Cobertura(object):
         """Return the version number of the coverage report."""
         return self.xml.attrib['version']
 
-    def line_rate(self, class_filename=None):
+    def line_rate(self, filename=None):
         """
         Return the global line rate of the coverage report. If the
-        `class_filename` is given, return the line rate of the class file.
+        `filename` file is given, return the line rate of the file.
         """
-        if class_filename is None:
+        if filename is None:
             el = self.xml
         else:
-            el = self._get_element_by_class_filename(class_filename)
+            el = self._get_class_element_by_filename(filename)
 
         return float(el.attrib['line-rate'])
 
-    def branch_rate(self, class_filename=None):
+    def branch_rate(self, filename=None):
         """
         Return the global branch rate of the coverage report. If the
-        `class_filename` is given, return the branch rate of the class file.
+        `filename` file is given, return the branch rate of the file.
         """
-        if class_filename is None:
+        if filename is None:
             el = self.xml
         else:
-            el = self._get_element_by_class_filename(class_filename)
+            el = self._get_class_element_by_filename(filename)
 
         return float(el.attrib['branch-rate'])
 
     @memoize
-    def missed_statements(self, class_filename):
+    def missed_statements(self, filename):
         """
         Return a list of uncovered line numbers for each of the missed
-        statements found for the class file `class_filename`.
+        statements found for the file `filename`.
         """
-        el = self._get_element_by_class_filename(class_filename)
+        el = self._get_class_element_by_filename(filename)
         lines = el.xpath('./lines/line[@hits=0]')
         return [int(l.attrib['number']) for l in lines]
 
     @memoize
-    def hit_statements(self, class_filename):
+    def hit_statements(self, filename):
         """
         Return a list of covered line numbers for each of the hit statements
-        found for the class file `class_filename`.
+        found for the file `filename`.
         """
-        el = self._get_element_by_class_filename(class_filename)
+        el = self._get_class_element_by_filename(filename)
         lines = el.xpath('./lines/line[@hits>0]')
         return [int(l.attrib['number']) for l in lines]
 
-    def line_statuses(self, class_filename):
+    def line_statuses(self, filename):
         """
         Return a list of tuples `(lineno, status)` of all the lines found in
-        the Cobertura report for the given class file where `lineno` is the
-        line number and `status` is coverage status of the line which can be
-        either `True` (line hit) or `False` (line miss).
+        the Cobertura report for the given file `filename` where `lineno` is
+        the line number and `status` is coverage status of the line which can
+        be either `True` (line hit) or `False` (line miss).
         """
-        line_elements = self._get_lines_by_class_filename(class_filename)
+        line_elements = self._get_lines_by_filename(filename)
 
         lines_w_status = []
         for line in line_elements:
@@ -123,29 +123,29 @@ class Cobertura(object):
 
         return lines_w_status
 
-    def missed_lines(self, class_filename):
+    def missed_lines(self, filename):
         """
-        Return a list of extrapolated uncovered line numbers for the class
-        file `class_filename` according to `Cobertura.line_statuses`.
+        Return a list of extrapolated uncovered line numbers for the
+        file `filename` according to `Cobertura.line_statuses`.
         """
-        statuses = self.line_statuses(class_filename)
+        statuses = self.line_statuses(filename)
         statuses = extrapolate_coverage(statuses)
         return [lno for lno, status in statuses if status is False]
 
     @memoize
-    def class_file_source(self, class_filename):
+    def file_source(self, filename):
         """
         Return a list of namedtuple `Line` for each line of code found in the
-        source file with the given `class_filename`.
+        source file with the given `filename`.
         """
-        filename = self.filepath(class_filename)
+        filepath = self.filepath(filename)
 
-        if not os.path.exists(filename):
-            return [Line(0, '%s not found' % filename, None, None)]
+        if not os.path.exists(filepath):
+            return [Line(0, '%s not found' % filepath, None, None)]
 
-        with codecs.open(filename, encoding='utf-8') as f:
+        with codecs.open(filepath, encoding='utf-8') as f:
             lines = []
-            line_statuses = dict(self.line_statuses(class_filename))
+            line_statuses = dict(self.line_statuses(filename))
             for lineno, source in enumerate(f, start=1):
                 line_status = line_statuses.get(lineno)
                 line = Line(lineno, source, line_status, None)
@@ -153,93 +153,83 @@ class Cobertura(object):
 
         return lines
 
-    def total_misses(self, class_filename=None):
+    def total_misses(self, filename=None):
         """
-        Return the total number of uncovered statements for the class file
-        `class_filename`. If `class_filename` is not given, return the total
-        number of uncovered statements for all class files.
+        Return the total number of uncovered statements for the file
+        `filename`. If `filename` is not given, return the total
+        number of uncovered statements for all files.
         """
-        if class_filename is not None:
-            return len(self.missed_statements(class_filename))
+        if filename is not None:
+            return len(self.missed_statements(filename))
 
         total = 0
-        for class_filename in self.class_files():
-            total += len(self.missed_statements(class_filename))
+        for filename in self.files():
+            total += len(self.missed_statements(filename))
 
         return total
 
-    def total_hits(self, class_filename=None):
+    def total_hits(self, filename=None):
         """
-        Return the total number of covered statements for the class file
-        `class_filename`. If `class_filename` is not given, return the total
-        number of covered statements for all class files.
+        Return the total number of covered statements for the file
+        `filename`. If `filename` is not given, return the total
+        number of covered statements for all files.
         """
-        if class_filename is not None:
-            return len(self.hit_statements(class_filename))
+        if filename is not None:
+            return len(self.hit_statements(filename))
 
         total = 0
-        for class_filename in self.class_files():
-            total += len(self.hit_statements(class_filename))
+        for filename in self.files():
+            total += len(self.hit_statements(filename))
 
         return total
 
-    def total_statements(self, class_filename=None):
+    def total_statements(self, filename=None):
         """
-        Return the total number of statements for the class file
-        `class_filename`. If `class_filename` is not given, return the total
-        number of statements for all class files.
+        Return the total number of statements for the file
+        `filename`. If `filename` is not given, return the total
+        number of statements for all files.
         """
-        if class_filename is not None:
-            statements = self._get_lines_by_class_filename(class_filename)
+        if filename is not None:
+            statements = self._get_lines_by_filename(filename)
             return len(statements)
 
         total = 0
-        for class_filename in self.class_files():
-            statements = self._get_lines_by_class_filename(class_filename)
+        for filename in self.files():
+            statements = self._get_lines_by_filename(filename)
             total += len(statements)
 
         return total
 
-    @memoize
-    def filename(self, class_filename):
+    def filepath(self, filename):
         """
-        Return the filename of the class file `class_filename` as found in the
-        Cobertura report.
-        """
-        el = self._get_element_by_class_filename(class_filename)
-        filename = el.attrib['filename']
-        return filename
-
-    def filepath(self, class_filename):
-        """
-        Return the filesystem path to the actual class file. It uses the
+        Return the filesystem path to the actual file `filename`. It uses the
         `base_path` value initialized in the constructor by prefixing it to the
-        class filename using `os.path.join(base_path, filename)`.
+        filename using `os.path.join(base_path, filename)`.
         """
-        filepath = os.path.join(self.base_path, class_filename)
+        filepath = os.path.join(self.base_path, filename)
         return filepath
 
     @memoize
-    def class_files(self):
+    def files(self):
         """
-        Return the list of available class files in the coverage report.
+        Return the list of available files in the coverage report.
         """
         return [el.attrib['filename'] for el in self.xml.xpath("//class")]
 
-    def has_classfile(self, class_filename):
+    def has_file(self, filename):
         """
-        Return `True` if the class file `class_filename` is present in the
-        report, return `False` otherwise.
+        Return `True` if the file `filename` is present in the report, return
+        `False` otherwise.
         """
         # FIXME: this will lookup a list which is slow, make it O(1)
-        return class_filename in self.class_files()
+        return filename in self.files()
 
     @memoize
-    def source_lines(self, class_filename):
+    def source_lines(self, filename):
         """
-        Return a list for source lines of class file `class_filename`.
+        Return a list for source lines of file `filename`.
         """
-        with codecs.open(self.filepath(class_filename), encoding='utf-8') as f:
+        with codecs.open(self.filepath(filename), encoding='utf-8') as f:
             return f.readlines()
 
     @memoize
@@ -265,8 +255,8 @@ class CoberturaDiff(object):
         This does not ensure that all changes have been covered. If this is
         what you want, use `CoberturaDiff.has_all_changes_covered()` instead.
         """
-        for class_filename in self.class_files():
-            if self.diff_total_misses(class_filename) > 0:
+        for filename in self.files():
+            if self.diff_total_misses(filename) > 0:
                 return False
         return True
 
@@ -274,8 +264,8 @@ class CoberturaDiff(object):
         """
         Return `True` if all changes have been covered, `False` otherwise.
         """
-        for class_name in self.class_files():
-            for hunk in self.class_source_hunks(class_name):
+        for filename in self.files():
+            for hunk in self.file_source_hunks(filename):
                 for line in hunk:
                     if line.reason is None:
                         continue  # line untouched
@@ -283,84 +273,84 @@ class CoberturaDiff(object):
                         return False  # line not covered
         return True
 
-    def _diff_attr(self, attr_name, class_filename):
+    def _diff_attr(self, attr_name, filename):
         """
         Return the difference between
-        `self.cobertura2.<attr_name>(class_filename)` and
-        `self.cobertura1.<attr_name>(class_filename)`.
+        `self.cobertura2.<attr_name>(filename)` and
+        `self.cobertura1.<attr_name>(filename)`.
 
         This generic method is meant to diff the count of methods that return
-        counts for a given class name, e.g. `Cobertura.total_statements`,
+        counts for a given file `filename`, e.g. `Cobertura.total_statements`,
         `Cobertura.total_misses`, ...
 
         The returned count may be a float.
         """
-        if class_filename is not None:
-            class_files = [class_filename]
+        if filename is not None:
+            files = [filename]
         else:
-            class_files = self.class_files()
+            files = self.files()
 
         total_count = 0.0
-        for class_filename in class_files:
-            if self.cobertura1.has_classfile(class_filename):
+        for filename in files:
+            if self.cobertura1.has_file(filename):
                 method = getattr(self.cobertura1, attr_name)
-                count1 = method(class_filename)
+                count1 = method(filename)
             else:
                 count1 = 0.0
             method = getattr(self.cobertura2, attr_name)
-            count2 = method(class_filename)
+            count2 = method(filename)
             total_count += count2 - count1
 
         return total_count
 
-    def diff_total_statements(self, class_filename=None):
-        return int(self._diff_attr('total_statements', class_filename))
+    def diff_total_statements(self, filename=None):
+        return int(self._diff_attr('total_statements', filename))
 
-    def diff_total_misses(self, class_filename=None):
-        return int(self._diff_attr('total_misses', class_filename))
+    def diff_total_misses(self, filename=None):
+        return int(self._diff_attr('total_misses', filename))
 
-    def diff_total_hits(self, class_filename=None):
-        return int(self._diff_attr('total_hits', class_filename))
+    def diff_total_hits(self, filename=None):
+        return int(self._diff_attr('total_hits', filename))
 
-    def diff_line_rate(self, class_filename=None):
-        return self._diff_attr('line_rate', class_filename)
+    def diff_line_rate(self, filename=None):
+        return self._diff_attr('line_rate', filename)
 
-    def diff_missed_lines(self, class_filename):
+    def diff_missed_lines(self, filename):
         """
         Return a list of 2-element tuples `(lineno, is_new)` for the given
-        class file `class_filename` where `lineno` is a missed line number and
-        `is_new` indicates whether the missed line was introduced (True) or
-        removed (False).
+        file `filename` where `lineno` is a missed line number and `is_new`
+        indicates whether the missed line was introduced (True) or removed
+        (False).
         """
         line_changed = []
-        for line in self.class_file_source(class_filename):
+        for line in self.file_source(filename):
             if line.status is not None:
                 is_new = not line.status
                 line_changed.append((line.number, is_new))
         return line_changed
 
-    def class_files(self):
+    def files(self):
         """
-        Return `self.cobertura2.class_files()`.
+        Return `self.cobertura2.files()`.
         """
-        return self.cobertura2.class_files()
+        return self.cobertura2.files()
 
-    def class_file_source(self, class_filename):
+    def file_source(self, filename):
         """
         Return a list of namedtuple `Line` for each line of code found in the
-        given class file `class_filename`.
+        given file `filename`.
 
         """
-        if self.cobertura1.has_classfile(class_filename):
-            lines1 = self.cobertura1.source_lines(class_filename)
+        if self.cobertura1.has_file(filename):
+            lines1 = self.cobertura1.source_lines(filename)
             line_statuses1 = dict(self.cobertura1.line_statuses(
-                class_filename))
+                filename))
         else:
             lines1 = []
             line_statuses1 = {}
 
-        lines2 = self.cobertura2.source_lines(class_filename)
-        line_statuses2 = dict(self.cobertura2.line_statuses(class_filename))
+        lines2 = self.cobertura2.source_lines(filename)
+        line_statuses2 = dict(self.cobertura2.line_statuses(filename))
 
         # Build a dict of lineno2 -> lineno1
         lineno_map = reconcile_lines(lines2, lines1)
@@ -392,13 +382,13 @@ class CoberturaDiff(object):
 
         return lines
 
-    def class_source_hunks(self, class_filename):
+    def file_source_hunks(self, filename):
         """
-        Like `CoberturaDiff.class_file_source`, but returns a list of line
-        hunks of the lines that have changed for the given `class_filename`.
-        An empty list means that the class has no lines that have a change in
-        coverage status.
+        Like `CoberturaDiff.file_source`, but returns a list of line hunks of
+        the lines that have changed for the given file `filename`. An empty
+        list means that the file has no lines that have a change in coverage
+        status.
         """
-        lines = self.class_file_source(class_filename)
+        lines = self.file_source(filename)
         hunks = hunkify_lines(lines)
         return hunks