+++ tests/test_cobertura.py
@@ -34,9 +34,9 @@ def test_line_rate_by_class_file():
         'search/LinearSearch.java': 0.7142857142857143,
     }
 
-    for class_filename in cobertura.class_files():
-        assert cobertura.line_rate(class_filename) == \
-            expected_line_rates[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.line_rate(filename) == \
+            expected_line_rates[filename]
 
 
 def test_branch_rate():
@@ -53,9 +53,9 @@ def test_branch_rate_by_class_file():
         'search/LinearSearch.java': 0.6666666666666666,
     }
 
-    for class_filename in cobertura.class_files():
-        assert cobertura.branch_rate(class_filename) == \
-            expected_branch_rates[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.branch_rate(filename) == \
+            expected_branch_rates[filename]
 
 
 def test_missed_statements_by_class_file():
@@ -67,9 +67,9 @@ def test_missed_statements_by_class_file():
         'search/LinearSearch.java': [19, 24],
     }
 
-    for class_filename in cobertura.class_files():
-        assert cobertura.missed_statements(class_filename) == \
-            expected_missed_statements[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.missed_statements(filename) == \
+            expected_missed_statements[filename]
 
 
 def test_list_packages():
@@ -82,7 +82,7 @@ def test_list_packages():
 def test_list_classes():
     cobertura = make_cobertura()
 
-    classes = cobertura.class_files()
+    classes = cobertura.files()
     assert classes == [
         'Main.java',
         'search/BinarySearch.java',
@@ -101,8 +101,8 @@ def test_hit_lines__by_iterating_over_classes():
         'search/LinearSearch.java': [9, 13, 15, 16, 17],
     }
 
-    for class_filename in cobertura.class_files():
-        assert cobertura.hit_statements(class_filename) == expected_lines[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.hit_statements(filename) == expected_lines[filename]
 
 
 def test_missed_lines():
@@ -115,8 +115,8 @@ def test_missed_lines():
         'search/LinearSearch.java': [19, 20, 21, 22, 23, 24],
     }
 
-    for class_filename in cobertura.class_files():
-        assert cobertura.missed_lines(class_filename) == expected_lines[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.missed_lines(filename) == expected_lines[filename]
 
 
 def test_total_statements():
@@ -132,9 +132,9 @@ def test_total_statements_by_class_file():
         'search/ISortedArraySearch.java': 0,
         'search/LinearSearch.java': 7,
     }
-    for class_filename in cobertura.class_files():
-        assert cobertura.total_statements(class_filename) == \
-            expected_total_statements[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.total_statements(filename) == \
+            expected_total_statements[filename]
 
 
 def test_total_misses():
@@ -150,9 +150,9 @@ def test_total_misses_by_class_file():
         'search/ISortedArraySearch.java': 0,
         'search/LinearSearch.java': 2,
     }
-    for class_filename in cobertura.class_files():
-        assert cobertura.total_misses(class_filename) == \
-            expected_total_misses[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.total_misses(filename) == \
+            expected_total_misses[filename]
 
 
 def test_total_hits():
@@ -168,9 +168,9 @@ def test_total_hits_by_class_file():
         'search/ISortedArraySearch.java': 0,
         'search/LinearSearch.java': 5,
     }
-    for class_filename in cobertura.class_files():
-        assert cobertura.total_hits(class_filename) == \
-            expected_total_misses[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.total_hits(filename) == \
+            expected_total_misses[filename]
 
 
 def test_filepath():
@@ -182,9 +182,9 @@ def test_filepath():
         'search/ISortedArraySearch.java': 'foo/bar/baz/search/ISortedArraySearch.java',
         'search/LinearSearch.java': 'foo/bar/baz/search/LinearSearch.java',
     }
-    for class_filename in cobertura.class_files():
-        assert cobertura.filepath(class_filename) == \
-            expected_filepaths[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.filepath(filename) == \
+            expected_filepaths[filename]
 
 
 def test_class_file_source__sources_not_found():
@@ -196,8 +196,8 @@ def test_class_file_source__sources_not_found():
         'search/ISortedArraySearch.java': [Line(0, 'tests/search/ISortedArraySearch.java not found', None, None)],
         'search/LinearSearch.java': [Line(0, 'tests/search/LinearSearch.java not found', None, None)],
     }
-    for class_filename in cobertura.class_files():
-        assert cobertura.class_file_source(class_filename) == expected_sources[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.file_source(filename) == expected_sources[filename]
 
 
 def test_line_statuses():
@@ -223,9 +223,9 @@ def test_line_statuses():
             (6, False)
         ],
     }
-    for class_filename in cobertura.class_files():
-        assert cobertura.line_statuses(class_filename) == \
-            expected_line_statuses[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.line_statuses(filename) == \
+            expected_line_statuses[filename]
 
 
 def test_class_file_source__sources_found():
@@ -254,6 +254,6 @@ def test_class_file_source__sources_found():
             Line(6, '    pass\n', False, None)
         ],
     }
-    for class_filename in cobertura.class_files():
-        assert cobertura.class_file_source(class_filename) == \
-               expected_sources[class_filename]
+    for filename in cobertura.files():
+        assert cobertura.file_source(filename) == \
+               expected_sources[filename]